spell с decimal

+++ Morpher.WebApi/General/Data/Interfaces/IRussianAnalyzer.cs
@@ -9,7 +9,7 @@
     {
         DeclensionResult Declension(string s, DeclensionFlags? flags = null);
 
-        NumberSpelling Spell(int n, string unit);
+        NumberSpelling Spell(decimal n, string unit);
 
         AdjectiveGenders AdjectiveGenders(string s);
